requirements.txt
# Web框架
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
pydantic>=2.3.0
starlette>=0.27.0
python-multipart>=0.0.6
httpx>=0.24.1

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# 文档处理
PyPDF2>=3.0.0
docx2txt>=0.8
beautifulsoup4>=4.12.0
markdown>=3.4.0
html2text>=2020.1.16
pytesseract>=0.3.10  # 如果需要OCR功能

# NLP & 向量处理
transformers>=4.32.0
sentence-transformers>=2.2.2
nltk>=3.8.1
spacy>=3.6.0
scikit-learn>=1.3.0

# 向量数据库客户端 (根据需要选择)
chromadb>=0.4.15  # Chroma
pymilvus>=2.3.0  # Milvus
pinecone-client>=2.2.2  # Pinecone
weaviate-client>=3.24.1  # Weaviate

# LLM客户端 (根据需要选择)
openai>=1.3.0  # OpenAI API
anthropic>=0.5.0  # Anthropic/Claude API

# 本地模型支持 (可选)
torch>=2.0.0
accelerate>=0.22.0
bitsandbytes>=0.41.0  # 用于量化

# 工具和工具性库
tqdm>=4.66.1
tenacity>=8.2.3  # 重试机制
tiktoken>=0.5.1  # OpenAI tokenizer
loguru>=0.7.0  # 更好的日志
cachetools>=5.3.1

# 开发工具 (可选)
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0 